Extract VaultHeader
