+++ features/manageVault/ManageVaultView.tsx
@@ -1,14 +1,13 @@
-import { Icon } from '@makerdao/dai-ui-icons'
 import { trackingEvents } from 'analytics/analytics'
 import BigNumber from 'bignumber.js'
-import { getToken } from 'blockchain/tokensMetadata'
 import { useAppContext } from 'components/AppContextProvider'
+import { VaultHeader } from 'components/vault/VaultHeader'
 import { ManageVaultFormHeader } from 'features/manageVault/ManageVaultFormHeader'
 import { AppSpinner, WithLoadingIndicator } from 'helpers/AppSpinner'
 import { useObservableWithError } from 'helpers/observableHook'
 import { useTranslation } from 'next-i18next'
 import React, { useEffect } from 'react'
-import { Box, Card, Divider, Flex, Grid, Heading, SxProps, Text } from 'theme-ui'
+import { Box, Card, Divider, Grid } from 'theme-ui'
 import { slideInAnimation } from 'theme/animations'
 
 import { ManageVaultState } from './manageVault'
@@ -64,39 +63,25 @@ function ManageVaultForm(props: ManageVaultState) {
   )
 }
 
-export function ManageVaultHeading(props: ManageVaultState & SxProps) {
-  const {
-    vault: { id, ilk, token },
-    sx,
-  } = props
-  const tokenInfo = getToken(token)
-  const { t } = useTranslation()
-  return (
-    <Heading
-      as="h1"
-      variant="paragraph2"
-      sx={{ gridColumn: ['1', '1/3'], fontWeight: 'semiBold', borderBottom: 'light', pb: 3, ...sx }}
-    >
-      <Flex sx={{ justifyContent: ['center', 'left'] }}>
-        <Icon name={tokenInfo.iconCircle} size="26px" sx={{ verticalAlign: 'sub', mr: 2 }} />
-        <Text>{t('vault.header', { ilk, id })}</Text>
-      </Flex>
-    </Heading>
-  )
-}
-
 export function ManageVaultContainer(props: ManageVaultState) {
+  const {
+    vault: { id, ilk },
+  } = props
+  const { t } = useTranslation()
+
   return (
-    <Grid mt={4} columns={['1fr', '2fr minmax(380px, 1fr)']} gap={5}>
-      <ManageVaultHeading {...props} sx={{ display: ['block', 'none'] }} />
-      <Box mb={6} sx={{ order: [3, 1] }}>
-        <ManageVaultDetails {...props} />
-      </Box>
-      <Divider sx={{ display: ['block', 'none'], order: [2, 0] }} />
-      <Box sx={{ order: [1, 2] }}>
-        <ManageVaultForm {...props} />
-      </Box>
-    </Grid>
+    <>
+      <VaultHeader {...{ ...props, header: t('vault.header', { ilk, id }), id }} />
+      <Grid mt={4} columns={['1fr', '2fr minmax(380px, 1fr)']} gap={5}>
+        <Box mb={6} sx={{ order: [3, 1] }}>
+          <ManageVaultDetails {...props} />
+        </Box>
+        <Divider sx={{ display: ['block', 'none'], order: [2, 0] }} />
+        <Box sx={{ order: [1, 2] }}>
+          <ManageVaultForm {...props} />
+        </Box>
+      </Grid>
+    </>
   )
 }
