+++ helpers/types.ts
@@ -1,3 +1,6 @@
+import { ManageVaultState } from 'features/manageVault/manageVault'
+import { OpenMultiplyVaultState } from 'features/openMultiplyVault/openMultiplyVault'
+import { OpenVaultState } from 'features/openVault/openVault'
 import { TFunction } from 'next-i18next'
 import { ParsedUrlQuery } from 'querystring'
 
@@ -5,3 +8,4 @@ export type WithChildren = { children?: any }
 export type WithTranslation = { t: TFunction }
 export type WithQuery = { query?: ParsedUrlQuery }
 export type WithReadonlyAccount = { readonlyAccount?: boolean }
+export type CommonVaultState = OpenVaultState | ManageVaultState | OpenMultiplyVaultState
